Repackage duke files
