+++ src/main/java/duke/Main.java
@@ -2,6 +2,7 @@ package duke;
 
 import java.io.IOException;
 
+import duke.controller.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
