+++ src/main/java/duke/controller/MainWindow.java
@@ -1,5 +1,6 @@
-package duke;
+package duke.controller;
 
+import duke.Duke;
 import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
