+++ src/main/java/duke/storage/ArchiveStorage.java
@@ -1,29 +1,29 @@
-package duke;
+package duke.storage;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.io.PrintWriter;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.List;
 
-/**
- * This class encapsulates the logic for saving and writing to the hard disk.
- */
-public class MainStorage extends Storage {
+import duke.util.Ui;
+
+public class ArchiveStorage extends Storage {
     /**
-     * Constructor for a new MainStorage object.
+     * Constructor for a new ArchiveStorage object.
      *
      * @param filePath The filepath to the save file for the application.
      */
-    public MainStorage(String filePath) {
+    public ArchiveStorage(String filePath) {
         super(filePath);
     }
 
     /**
-     * Saves the input string to the instance file path.
+     * Appends and saves the input string to the instance file path.
      *
      * @param text The text to be saved.
      */
@@ -35,25 +35,30 @@ public class MainStorage extends Storage {
             // Create parent directories if they do not exist
             file.getParentFile().mkdirs();
 
-            FileWriter fw = new FileWriter(this.filePath);
+            // Append to file
+            FileWriter fw = new FileWriter(this.filePath, true);
             fw.write(text);
             fw.close();
         } catch (IOException e) {
-            Ui.printMessage(String.format("An error occurred while saving to '%s'.", this.filePath));
+            Ui.printMessage(String.format("An error occurred while saving to archive file '%s'.", this.filePath));
         } catch (SecurityException e) {
-            Ui.printMessage(String.format("Unable to write to '%s'. Please check your directory and file permissions.",
+            Ui.printMessage(String.format("Unable to write to archive file '%s'. "
+                            + "Please check your directory and file permissions.",
                     this.filePath));
         }
     }
 
     /**
      * Loads the file specified by the instance file path.
+     * Deletes the file contents after loading.
      *
      * @return The file contents as a list of strings.
      * @throws FileNotFoundException When file is not found at the specified path.
      */
     @Override
     public List<String> load() throws IOException {
-        return Files.readAllLines(Paths.get(this.filePath), Charset.defaultCharset());
+        List<String> fileContents = Files.readAllLines(Paths.get(this.filePath), Charset.defaultCharset());
+        new PrintWriter(this.filePath).close(); // Delete file contents after loading archive file
+        return fileContents;
     }
 }
