+++ src/main/java/duke/util/Parser.java
@@ -1,9 +1,14 @@
-package duke;
+package duke.util;
 
 import java.io.IOException;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+
+import duke.TaskExecution;
+import duke.exception.DukeException;
+import duke.exception.InvalidFormatException;
+import duke.exception.UnknownCommandException;
 
 /**
  * This class encapsulates the parsing logic for the application.
