+++ src/test/java/duke/TaskListTest.java
@@ -7,6 +7,12 @@ import java.util.ArrayList;
 
 import org.junit.jupiter.api.Test;
 
+import duke.storage.Event;
+import duke.task.Deadline;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.task.Todo;
+
 public class TaskListTest {
     @Test
     void constructor_emptyInput_success() {
